# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Movies - API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/medias": {
            "get": {
                "tags": [
                    "Media"
                ],
                "summary": "Gets the list of media",
                "description": "Returns the list of media with optional pagination",
                "operationId": "getMediaList",
                "parameters": [
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Number of medias per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No media(s) found"
                    }
                }
            }
        },
        "/api/medias/{id}": {
            "get": {
                "tags": [
                    "Media"
                ],
                "summary": "Get media by ID",
                "description": "Returns a single media",
                "operationId": "getMediaById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of media to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "format": "int64",
                                            "example": 1
                                        },
                                        "media_path": {
                                            "type": "string",
                                            "example": "/path/to/media"
                                        },
                                        "media_url": {
                                            "type": "string",
                                            "example": "http://example.com/media"
                                        },
                                        "media_type": {
                                            "type": "string",
                                            "example": "image"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2021-09-15T00:00:00Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2021-09-15T00:00:00Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Media not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Media"
                ],
                "summary": "Delete a media",
                "description": "Deletes a media and returns a success message",
                "operationId": "deleteMedia",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of media to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Media deleted successfully"
                    },
                    "404": {
                        "description": "Media not found"
                    },
                    "500": {
                        "description": "Error deleting media"
                    }
                }
            }
        },
...
    }
}